requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
celery==5.4.0
redis==5.1.0
python-multipart==0.0.9
pydantic==2.9.2
python-jose[cryptography]==3.3.0
Pillow==10.4.0
pytesseract==0.3.13
